fix context connection problem by making the form tree a component
