+++ frontend/src/components/applyForm/PrintForm.tsx
@@ -2,11 +2,8 @@
 
 import { RJSFSchema } from "@rjsf/utils";
 
-import { JSX } from "react";
-import { Alert } from "@trussworks/react-uswds";
-
+import { FormFields } from "./FormFields";
 import { UiSchema } from "./types";
-import { buildFormTreeRecursive } from "./utils";
 
 export default function PrintForm({
   formSchema,
@@ -17,22 +14,12 @@ export default function PrintForm({
   savedFormData: object;
   uiSchema: UiSchema;
 }) {
-  let fields: JSX.Element[] = [];
-  try {
-    fields = buildFormTreeRecursive({
-      errors: null,
-      formData: savedFormData,
-      schema: formSchema,
-      uiSchema,
-    });
-  } catch (e) {
-    console.error(e);
-    return (
-      <Alert data-testid="alert" type="error" heading="Error" headingLevel="h4">
-        Error rendering form
-      </Alert>
-    );
-  }
-
-  return <>{fields}</>;
+  return (
+    <FormFields
+      errors={null}
+      formData={savedFormData}
+      schema={formSchema}
+      uiSchema={uiSchema}
+    />
+  );
 }
